通信のア処理はとりあえず完成
クライアント側一応完成．あとは環境に合わせてパラメタ調整するのみ．通信は，文字列に格納しなくてもArray[Byte]で上手くデータを格納できた．

+++ src/main/scala/Main.scala
@@ -40,7 +40,7 @@ object Main extends processing.core.PApplet{
 	override def sketchFullScreen():Boolean = {
   		return false
 	}
-	def receive(data:Array[Int]){
+	def receive(data:Array[Byte]){
 		network.receive(data)
 	}
 }
